+++ packages/terra-form-select/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+6.0.0 - (March 31, 2020)
+------------------
+### Breaking Changes
+* `label` prop's type changed from node to string.
+
+5.45.0 - (March 31, 2020)
+------------------
+### Changed
+* Fixed `onBlur` event for mobile to not be triggered when selecting an option
+
 5.44.0 - (March 10, 2020)
 ------------------
 ### Changed