+++ packages/terra-i18n/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+4.26.0 - (March 31, 2020)
+------------------
+### Fixed
+* Check if Intl.DateTimeFormat and Int.NumberFormat have the `supportedLocalesOf` property before adding as constructors for intl-locales-supported. 
+
 4.25.0 - (March 10, 2020)
 ------------------
 ### Changed