+++ packages/terra-search-field/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+3.43.0 - (March 31, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.42.0 - (March 31, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.41.0 - (March 10, 2020)
 ------------------
 ### Changed