+++ packages/terra-status-view/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+4.20.0 - (March 31, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 4.19.0 - (March 10, 2020)
 ------------------
 ### Changed